\name{Factors}
\alias{Factors}
\docType{data}
\title{
FORAM Index Factors
}
\description{
An example dataset defining factors levels to use with the \code{\link[forams]{fi}} function.
}
\usage{data(Factors)}
\format{
  A data frame with 29 observations on the following variable.
  \describe{
    \item{FI}{a factor with levels: \code{Ph}, \code{Po} and \code{Ps}}
  }
}
\details{
This dataset is an artificial random generated example. Unfortunately at the moment, due to authorship issues I can not present any real one. This may change in future versions.
}
\examples{
data(Factors)
summary(Factors)
}
\keyword{datasets}
