\name{fi-class}
\Rdversion{1.1}
\docType{class}
\alias{fi-class}

\title{Class \code{"fi"}}
\description{
Class used to store \code{"fi"} analysis objects.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("fi", ...)}.
Single slot classe used to store a \code{"data.frame"} object.
}
\section{Slots}{
  \describe{
    \item{\code{fi}:}{Object of class \code{"data.frame"} ~~ }
  }
}
\section{Methods}{
  \describe{
    \item{plot}{\code{signature(x = "fi")}: ... }
	 }
}
\author{
Rodrigo Aluizio
}
\seealso{
See Also as \code{\link[forams]{fi}}}

\examples{
showClass("fi")
}
\keyword{classes}
