\name{LF}
\alias{LF}
\docType{data}
\title{
FORAM Index and SHE dataset
}
\description{
An example dataset containing some \emph{taxa} abundances to use in exemples of the \code{\link[forams]{fi}} and \code{\link[forams]{she}} functions.
}
\usage{data(LF)}
\format{
  A data frame with 29 \emph{taxa} on 23 sites.
}
\details{
This dataset is an artificial random generated example. Unfortunately at the moment, due to authorship issues I can not present any real one. This may change in future versions.
}
\examples{
data(LF)
str(LF)
}
\keyword{datasets}
