\name{she-class}
\Rdversion{1.1}
\docType{class}
\alias{she-class}

\title{Class \code{"she"}}
\description{
Class used to store \code{"she"} analysis objects.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("she", ...)}.
Single slot classe used to store a \code{"data.frame"} object.
}
\section{Slots}{
  \describe{
    \item{\code{bi}:}{Object of class \code{"data.frame"} ~~ }
  }
}
\section{Methods}{
  \describe{
    \item{plot}{\code{signature(x = "she")}: ... }
	 }
}

\author{
Rodrigo Aluizio
}

\seealso{
See Also as \code{\link[forams]{she}}}

\examples{
showClass("she")
}
\keyword{classes}
